stream_pose_ml/requirements.txt
Flask==2.2.2
Flask_Cors==3.0.10
Flask_SocketIO==5.3.3
Werkzeug==2.2.2
eventlet==0.33.0
dnspython==2.2.1
imbalanced_learn==0.10.1
kneed==0.8.5
matplotlib==3.7.0
mediapipe>=0.10.5
numpy==1.24.2
opencv_contrib_python==4.7.0.68
pandas==1.5.3
python_engineio==4.4.0
PyWavelets==1.4.1
PyYAML
scikit_learn==1.2.1
scipy==1.11.2
seaborn==0.12.2
tqdm==4.65.0
tslearn==0.6.2
xgboost==1.7.4
pytest
pytest-emoji
pytest-cov
mlflow
